Update PubSub_SWG_Charter.adoc
Editorial changes before moving the charter to pending documents.

I think we also decided in the web-meeting to remove the background section at the end as it does not add value to the current work and may be more confusing. All the references that are relevant for the new charter are referenced in the charter.